+++ R/plot_flux.R
@@ -1,12 +1,13 @@
-#' Plot net bird movement
+#' Plot bird flux
 #'
-#' @param flux A flux object as created by [calc_flux(format = "dataframe")][calc_flux]
+#' @param flux A flux object as created by
+#'   [calc_flux(format = "dataframe")][calc_flux]
 #' @param bf A BirdFlow object
 #' @param subset A subset of the transitions in `flux` to plot, can be
 #' a logical vector of the same length as the number of transitions in `flux`;
 #' a numeric index of transitions in `flux`, or a subset of the transition names
 #' in `flux`.
-#' @param limits Two numbers representing the range in movement values to
+#' @param limits Two numbers representing the range in flux values to
 #' display. Values outside of this range will be truncated to the range. With
 #' the default of `NULL` the entire range is plotted.
 #' @param dynamic_scale  If `TRUE` then the range of the data in each
@@ -17,10 +18,10 @@
 #' to skip plotting the coastline.
 #' @param coast_color The color used to plot the coastline, or `NULL` to skip
 #' plotting the coastline.
-#' @param gradient_colors The colors palette used to represent the movement
+#' @param gradient_colors The colors palette used to represent the flux
 #' intensity.
 #' @param title The plot title
-#' @param value_label The label for the net movement values.
+#' @param value_label The label for the flux values.
 #' @return `plot_flux` returns a **ggplot2** object.  It can be displayed with
 #' `print()`.
 #' @export
@@ -34,7 +35,7 @@ plot_flux <- function(flux,
                       coast_color = gray(0.5),
                       gradient_colors = NULL,
                       title = species(bf),
-                      value_label = "Movement") {
+                      value_label = "Flux") {
 
   if (!is.null(limits) && dynamic_scale) {
     stop("Do not set dynamic_scale to TRUE while also setting limits.")
@@ -46,14 +47,14 @@ plot_flux <- function(flux,
 
 
   if (is.null(limits)) {
-    limits <- range(flux$movement, na.rm = TRUE)
+    limits <- range(flux$flux, na.rm = TRUE)
   } else {
     stopifnot(is.numeric(limits), length(limits) == 2, all(!is.na(limits)),
               limits[1] < limits[2])
     # Truncate to limits
 
-    flux$movement[flux$movement < limits[1]] <- limits[1]
-    flux$movement[flux$movement > limits[2]] <- limits[2]
+    flux$flux[flux$flux < limits[1]] <- limits[1]
+    flux$flux[flux$flux > limits[2]] <- limits[2]
   }
 
 
@@ -101,7 +102,7 @@ plot_flux <- function(flux,
   p <- flux |>
     #dplyr::filter(transition %in% transitions[seq(4, 50, 4)]) |>
     ggplot2::ggplot(ggplot2::aes(x = .data$x, y = .data$y,
-                                 fill = .data$movement)) +
+                                 fill = .data$flux)) +
     ggplot2::geom_raster() +
     ggplot2::scale_fill_gradientn(colors = gradient_colors)
 
@@ -111,11 +112,11 @@ plot_flux <- function(flux,
     # Multiple transitions, facet wrap on date and add species title
     p <- p +
       ggplot2::facet_wrap(ggplot2::vars(.data$date)) +
-      ggplot2::ggtitle(species(bf))
+      ggplot2::ggtitle(title)
   } else {
     # Single transition add species title AND date subtitle
     p <- p +
-      ggplot2::ggtitle(species(bf), subtitle = flux$date[1])
+      ggplot2::ggtitle(title, subtitle = flux$date[1])
   }
 
   # Add coastline